#301 fix circular import
