+++ src/artemis/external_interaction/exceptions.py
@@ -1,0 +1,4 @@
+class ISPyBDepositionNotMade(Exception):
+    """Raised when the ISPyB or Zocalo callbacks can't access ISPyB deposition numbers."""
+
+    pass
